+++ tests/gmprocess/subcommands/compute_station_metrics_test.py
@@ -34,7 +34,7 @@ def test_compute_station_metrics(script_runner):
         assert "Added station metrics to workspace files with" in ret.stderr
         assert "Calculating station metrics for CI.CCC.HN" in ret.stderr
 
-        ret = script_runner.run("gmrecords", "compute_station_metrics", "-o")
+        ret = script_runner.run("gmrecords", "-o", "compute_station_metrics")
         assert ret.success
 
     except Exception as ex: